// src/types.ts
export interface Episode {
    id: number;
    name: string;
    air_date: string;
    episode: string;
    characters: string[];
    created: string;
    url: string;
}
export interface Character {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: {
        name: string;
        url: string;
    };
    location: {
        name: string;
        url: string;
    };
    image: string;
    episode: string[];
    url: string;
    created: string;
}
export interface Location {
    id: number;
    name: string;
    type: string;
    dimension: string;
    residents: string[];
    location: object;
    url: string;
    created: string;
}